Hindi Language added to Nearby Search
