+++ lib/widgets/trackScreenDetailsWidget.dart
@@ -34,7 +34,6 @@ class TrackScreenDetails extends StatefulWidget {
   var imei;
   var recentStops;
 
-
   TrackScreenDetails({
     required this.gpsData,
     required this.gpsTruckRoute,
@@ -52,7 +51,6 @@ class TrackScreenDetails extends StatefulWidget {
     required this.totalDistance,
     this.imei,
     this.recentStops,
-
   });
 
   @override
@@ -80,8 +78,8 @@ class _TrackScreenDetailsState extends State<TrackScreenDetails> {
   void initState() {
     super.initState();
     initFunction();
-    timer = Timer.periodic(Duration(minutes: 0, seconds: 10),
-          (Timer t) => initFunction());
+    timer = Timer.periodic(
+        Duration(minutes: 0, seconds: 10), (Timer t) => initFunction());
   }
 
   initFunction() {
@@ -108,12 +106,13 @@ class _TrackScreenDetailsState extends State<TrackScreenDetails> {
       throw 'Could not open the map.';
     }
   }
+
   @override
   void dispose() {
-    
     timer.cancel();
     super.dispose();
   }
+
   @override
   Widget build(BuildContext context) {
     double height = MediaQuery.of(context).size.height;
@@ -255,51 +254,44 @@ class _TrackScreenDetailsState extends State<TrackScreenDetails> {
                         margin: EdgeInsets.fromLTRB(0, 0, 0, 0),
                         child: Column(
                           children: [
-                            Image.asset('assets/icons/speed_status.png',width:35,height: 35),
+                            Image.asset('assets/icons/speed_status.png',
+                                width: 35, height: 35),
                             (widget.gpsData.last.speed > 2)
                                 ? Row(
-                                  crossAxisAlignment: CrossAxisAlignment.center,
-                                  children: 
-                                  [Text(
-                                      "${(gpsData.last.speed).round()} " ,
-                                      style: TextStyle(
-                                          color: liveasyGreen,
-                                          fontSize: size_10,
-                                          fontStyle: FontStyle.normal,
-                                          fontWeight: regularWeight)),
-                                    Text(
-                                      
-                                          "km/h".tr,
-                                      style: TextStyle(
-                                          color: liveasyGreen,
-                                          fontSize: size_6,
-                                          fontStyle: FontStyle.normal,
-                                          fontWeight: regularWeight)),
-                                  ]
-                                )
-                                  
+                                    crossAxisAlignment:
+                                        CrossAxisAlignment.center,
+                                    children: [
+                                        Text("${(gpsData.last.speed).round()} ",
+                                            style: TextStyle(
+                                                color: liveasyGreen,
+                                                fontSize: size_10,
+                                                fontStyle: FontStyle.normal,
+                                                fontWeight: regularWeight)),
+                                        Text("km/h".tr,
+                                            style: TextStyle(
+                                                color: liveasyGreen,
+                                                fontSize: size_6,
+                                                fontStyle: FontStyle.normal,
+                                                fontWeight: regularWeight)),
+                                      ])
                                 : Row(
-                                  crossAxisAlignment: CrossAxisAlignment.center,
-                                  children: [
-                                    Text(
-                                      "${(gpsData.last.speed).round()} " ,
-                                         
-                                      style: TextStyle(
-                                          color: red,
-                                          fontSize: size_10,
-                                          fontStyle: FontStyle.normal,
-                                          fontWeight: regularWeight)),
-                                    Text(
-                                      
-                                          "km/h".tr,
-                                      style: TextStyle(
-                                          color: red,
-                                          fontSize: size_6,
-                                          fontStyle: FontStyle.normal,
-                                          fontWeight: regularWeight)),
-                                  ]
-                                ),
-                        /*    Text("status".tr,
+                                    crossAxisAlignment:
+                                        CrossAxisAlignment.center,
+                                    children: [
+                                        Text("${(gpsData.last.speed).round()} ",
+                                            style: TextStyle(
+                                                color: red,
+                                                fontSize: size_10,
+                                                fontStyle: FontStyle.normal,
+                                                fontWeight: regularWeight)),
+                                        Text("km/h".tr,
+                                            style: TextStyle(
+                                                color: red,
+                                                fontSize: size_6,
+                                                fontStyle: FontStyle.normal,
+                                                fontWeight: regularWeight)),
+                                      ]),
+                            /*    Text("status".tr,
                                 style: TextStyle(
                                     color: black,
                                     fontSize: size_6,
@@ -453,10 +445,13 @@ class _TrackScreenDetailsState extends State<TrackScreenDetails> {
                           heroTag: "button1",
                           backgroundColor: Colors.white,
                           foregroundColor: bidBackground,
-                          child: Image.asset('assets/icons/navigate2.png',
-                          scale: 2.5,),
+                          child: Image.asset(
+                            'assets/icons/navigate2.png',
+                            scale: 2.5,
+                          ),
                           onPressed: () {
-                            openMap(gpsData.last.latitude, gpsData.last.longitude);
+                            openMap(
+                                gpsData.last.latitude, gpsData.last.longitude);
                           },
                         ),
                       ),
@@ -503,7 +498,8 @@ class _TrackScreenDetailsState extends State<TrackScreenDetails> {
                           heroTag: "button3",
                           backgroundColor: Colors.white,
                           foregroundColor: bidBackground,
-                          child: const Icon(Icons.play_circle_outline, size: 30),
+                          child:
+                              const Icon(Icons.play_circle_outline, size: 30),
                           onPressed: () {
                             Get.to(PlayRouteHistory(
                               gpsTruckHistory: gpsDataHistory,
@@ -607,10 +603,7 @@ class _TrackScreenDetailsState extends State<TrackScreenDetails> {
                                 placeOnTheMapName: "petrol_pump".tr,
                                 // position: position,
                                 TruckNo: widget.TruckNo,
-                                driverName: widget.driverName,
-                                driverNum: widget.driverNum,
                                 truckId: widget.truckId,
-                                gpsDataHistory: widget.gpsDataHistory,
                               ),
                             );
                           },
@@ -654,10 +647,7 @@ class _TrackScreenDetailsState extends State<TrackScreenDetails> {
                                 placeOnTheMapName: "police_station".tr,
                                 // position: position,
                                 TruckNo: widget.TruckNo,
-                                driverName: widget.driverName,
-                                driverNum: widget.driverNum,
                                 truckId: widget.truckId,
-                                gpsDataHistory: widget.gpsDataHistory,
                               ),
                             );
                           },
@@ -694,12 +684,11 @@ class _TrackScreenDetailsState extends State<TrackScreenDetails> {
                           ),
                           onPressed: () {
                             Get.to(truckAnalysisScreen(
-                              recentStops: recentStops,
-                              truckId: widget.truckId,
-                              TruckNo: widget.TruckNo,
-                              imei: widget.imei,
-                              deviceId: widget.deviceId
-                            ));
+                                recentStops: recentStops,
+                                truckId: widget.truckId,
+                                TruckNo: widget.TruckNo,
+                                imei: widget.imei,
+                                deviceId: widget.deviceId));
                           },
                         ),
                       ),
